+++ NEWS.md
@@ -2,6 +2,12 @@
 
 ## Polars R Package (development version)
 
+### New features
+
+- `pl$read_ipc()` can read a raw vector of Apache Arrow IPC file (#1072).
+- New method `<DataFrame>$to_raw_ipc()` to serialize a DataFrame to a raw vector
+  of Apache Arrow IPC file format (#1072).
+
 ## Polars R Package 0.16.3
 
 ### New features